+++ clang/lib/Frontend/CompilerInvocation.cpp
@@ -1905,16 +1905,6 @@ bool CompilerInvocation::ParseCodeGenArgs(CodeGenOptions &Opts, ArgList &Args,
     Opts.setInlining(CodeGenOptions::NormalInlining);
   }
 
-  // -mcmodel option.
-  if (const llvm::opt::Arg *A =
-          Args.getLastArg(clang::driver::options::OPT_mcmodel_EQ)) {
-    llvm::StringRef modelName = A->getValue();
-    if (modelName == "tiny" && !(T.isARM() || T.isAArch64())) {
-      Diags.Report(diag::err_drv_unsupported_option_argument_for_target)
-          << A->getSpelling() << modelName << T.getTriple();
-    }
-  }
-
   // PIC defaults to -fno-direct-access-external-data while non-PIC defaults to
   // -fdirect-access-external-data.
   Opts.DirectAccessExternalData =